+++ docs/.vuepress/configs/sidebar/en.ts
@@ -4,8 +4,18 @@ export const sidebarEn: SidebarConfig = {
   '/': [
     {
       text: 'Notes',
+      link: '/',
+    },
+    {
+      text: 'Work',
+      link: '/work/',
+      collapsible: true,
+    },
+    {
+      text: 'Writing',
+      link: '/writing/',
+      collapsible: true,
       children: [
-        '/',
         '/writing/digital-garden.md',
         '/writing/static-website-vuepress-github-pages.md',
       ],