+++ docs/changelog.md
@@ -2,6 +2,21 @@
 lang: en-US
 title: Changelog
 ---
+## 2022-08-17 — New notes
+- **Added** 
+  - [Работа](/ru/work/) 🇷🇺
+  - [Поиск работы](/ru/work/job-search.md) 🇷🇺
+  - [Резюме](/ru/work/cv.md) 🇷🇺
+  - [Собеседование](/ru/work/interview/) 🇷🇺
+  - [Ответы на мелкие вопросы](/ru/work/interview/trivia.md) 🇷🇺
+  - [Coding интервью](/ru/work/interview/coding.md) 🇷🇺
+  - [System Design интервью](/ru/work/interview/system-design.md) 🇷🇺
+  - [Behavioral интервью](/ru/work/interview/behavioral.md) 🇷🇺
+  - [Выбор оффера](/ru/work/offer.md) 🇷🇺
+  
+- **Translated** 
+  - [My Notes](/) -> [Мои заметки](/ru/)
+
 ## 2022-08-09 — Migration to VuePress
 - Migrated to VuePress